Add branch_matching.xml
This manifest will be used to check if repo-resource set the revision
to the right branch.
Indeed when we execute git ls-remote, the pattern is interpreted as a
glob [1].
Thus we may retrieve the revision of several remote branchs and
repo-resource must set the revision to the right branch.

In the demo-manifests project, when we git ls-remote with 'rev'
pattern, we will get the revision of these branchs:
fixed-rev-1
fixed-rev-2
rev

[1] https://git-scm.com/docs/git-ls-remote/en#Documentation/git-ls-remote.txt-ltpatternsgt82308203

Signed-off-by: Julien Masson <jmasson@baylibre.com>